blockchain/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pycryptodome==3.10.1
requests==2.26.0
click==8.0.1
pytest==6.2.5
python-dotenv==0.19.0
web3==5.24.0
numpy==1.21.0
aiohttp==3.8.1
cryptography==3.4.7
py-ecc==5.2.0
eth-utils==1.10.0
eth-keys==0.3.3
eth-account==0.5.6